Added perl script
The problem with cross compiling for the raspberry pi are the absolute symbolic links to the libraries and the absolute paths in the ld scripts. (SO Files).

If you copy prebuild libraries from the raspberry pi into your cross compile project, then you have to fix that. It took me some time to understand the strange linker issues while working on my project.

The project works as follows:

sync_package.pl <debian package name>

The output is a tar file with relative links and relative pathes in the ld scripts.

Example:

sync_package.pl libmodbus-dev

Output:

sysroot-libmodbus-dev.tar